+++ phpunit.bootstrap.php
@@ -4,4 +4,4 @@ declare(strict_types=1);
 
 require __DIR__ . '/vendor/autoload.php';
 
-define('ASSET_DIR', __DIR__ . '/tests/assets');
+DG\BypassFinals::enable();